+++ examples/example.py
@@ -5,8 +5,6 @@ from matplotlib import pyplot as plt
 
 from roipoly import RoiPoly
 
-logger = logging.getLogger(__name__)
-
 logging.basicConfig(format='%(levelname)s ''%(processName)-10s : %(asctime)s '
                            '%(module)s.%(funcName)s:%(lineno)s %(message)s',
                     level=logging.INFO)